refactor(orchestrator): extract claude/codex backends to file-per-backend (Step 1)
Move the claudeBackend + codexBackend const definitions out of registry.ts
into their own sibling files (claude.ts, codex.ts), matching the
file-per-backend pattern already used by openclaw.ts and acp.ts. registry.ts
now imports them instead of defining them inline.

Zero behavior change — the moved object literals are byte-identical to the
originals (only the `export` keyword is added); registry.ts changes only its
imports + removed defs. This breaks the future moa.ts <-> registry.ts import
cycle (the Collide backend needs claudeBackend + codexBackend as proposers
without importing the registry that will register Collide).

Proof: const bodies diff-clean vs origin/main; tsc --noEmit clean; full
vitest suite 344/344 green.
